requirements.txt
fastapi
uvicorn
python-multipart
requests
chromadb
sentence-transformers
langchain
langchain-community
langchain-huggingface
pdfplumber
pymupdf
pytesseract
Pillow
python-dotenv
prometheus-client
pytest
httpx
